refactor: Update script to scrape PAN data from GST website
The code change modifies the script to scrape PAN data from the GST website. Previously, the script was scraping data from the first column of the worksheet, but now it has been updated to scrape data from the third column. This change ensures that the correct PAN numbers are retrieved and processed.